src/requirements.txt
boto3==1.34.74
python-dotenv==0.19.2